move file PlayerId from tournament registration domain to shared folder - player profile module also will need it :-)
