+++ backend/src/modules/tournaments-registrations/core/application/command/AvailablePlayersForTournament.ts
@@ -1,4 +1,4 @@
-import { PlayerId } from '../../domain/PlayerId';
+import { PlayerId } from '../../../../../shared/core/domain/PlayerId';
 
 export interface AvailablePlayersForTournament {
   canPlay(playerId: PlayerId): Promise<boolean>;
