+++ backend/src/modules/tournaments-registrations/core/application/command/Players.ts
@@ -1,4 +1,4 @@
-import { PlayerId } from '../../domain/PlayerId';
+import { PlayerId } from '../../../../../shared/core/domain/PlayerId';
 
 export type Player = { playerId: PlayerId };
