+++ backend/src/modules/tournaments-registrations/core/application/event/PlayerProfileWasCreatedEventHandler.ts
@@ -1,7 +1,7 @@
 import { EventHandler } from '../../../../../shared/core/application/event/EventHandler';
 import { PlayerProfileWasCreated } from '../../../../player-profiles/core/domain/event/PlayerProfileWasCreated';
 import { Players } from '../command/Players';
-import { PlayerId } from '../../domain/PlayerId';
+import { PlayerId } from '../../../../../shared/core/domain/PlayerId';
 
 export class PlayerProfileWasCreatedEventHandler implements EventHandler<PlayerProfileWasCreated> {
   constructor(private readonly players: Players) {}
